Raise for persistence casades set with viewonly=True
An error is raised if any persistence-related "cascade" settings are made
on a :func:`.relationship` that also sets up viewonly=True.   The "cascade"
settings now default to non-persistence related settings only when viewonly
is also set.  This is the continuation from :ticket:`4993` where this
setting was changed to emit a warning in 1.3.

Fixes: #4994
Change-Id: Ic70ff4d9980e422ade474c5a0ad49756c6b8a048